pyproject.toml
[project]
name = "kasumi-next"
version = "0.1.0"
description = "一个「キラキラドキドキ」的 BanGDream! 小游戏机器人"
authors = [
    {name = "Kasumi-Games", email = "2667292003@qq.com"},
]
dependencies = [
    "nonebot-adapter-satori>=0.12.2",
    "nonebot-plugin-waiter>=0.6.2",
    "bestdori-api>=1.1.2",
    "bestdori-render>=0.1.1",
    "fuzzywuzzy>=0.18.0",
    "Pillow>=10.3.0",
    "aiohttp>=3.9.5",
    "nonebot-plugin-apscheduler>=0.4.0",
    "setuptools>=70.0.0",
    "python-Levenshtein>=0.25.1",
    "sqlalchemy>=2.0.30",
    "nonebot-plugin-localstore>=0.6.0",
    "aiofiles>=23.2.1",
    "nonebot2[httpx,websockets]>=2.3.1",
    "cairosvg>=2.7.1",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
